+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.java
@@ -23,6 +23,7 @@ import at.hannibal2.skyhanni.features.event.diana.GriffinBurrowHelper;
 import at.hannibal2.skyhanni.features.event.diana.GriffinBurrowParticleFinder;
 import at.hannibal2.skyhanni.features.event.diana.SoopyGuessBurrow;
 import at.hannibal2.skyhanni.features.fishing.*;
+import at.hannibal2.skyhanni.features.garden.SkyMartBestProfit;
 import at.hannibal2.skyhanni.features.inventory.*;
 import at.hannibal2.skyhanni.features.itemabilities.FireVeilWandParticles;
 import at.hannibal2.skyhanni.features.itemabilities.abilitycooldown.ItemAbilityCooldown;
@@ -108,6 +109,7 @@ public class SkyHanniMod {
         loadModule(new ItemTipHelper());
         loadModule(new RenderLivingEntityHelper());
         loadModule(new SkillExperience());
+        loadModule(new InventoryData());
 
         //features
         loadModule(new BazaarOrderHelper());
@@ -194,6 +196,7 @@ public class SkyHanniMod {
         loadModule(new TpsCounter());
         loadModule(new ParticleHider());
         loadModule(new MiscFeatures());
+        loadModule(new SkyMartBestProfit());
 
         Commands.INSTANCE.init();