+++ src/main/java/at/hannibal2/skyhanni/events/InventoryOpenEvent.kt
@@ -1,0 +1,5 @@
+package at.hannibal2.skyhanni.events
+
+import at.hannibal2.skyhanni.data.InventoryData
+
+class InventoryOpenEvent(val inventory: InventoryData.Inventory): LorenzEvent()